pyproject.toml
[tool.poetry]
name = "gfbot"
version = "0.1.0"
description = ""
authors = [""]
readme = "README.md"


[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.94.1"
typer = {extras = ["all"], version = "^0.7.0"}
python-dotenv = "^1.0.0"
httpx = "^0.23.3"
uvicorn = {extras = ["standard"], version = "^0.21.1"}
loguru = "^0.6.0"
pycrypto = "^2.6.1"
lmdb = "^1.4.0"
gunicorn = "^20.1.0"


[tool.poetry.group.dev.dependencies]
black = "^23.1.0"
mypy = "^1.1.1"
isort = "^5.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
